backend/requirements.txt
asgiref==3.6.0
Django==4.1.7
django-ckeditor==6.5.1
django-cors-headers==3.14.0
django-filter==23.1
django-js-asset==2.0.0
django-mysql==4.9.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Markdown==3.4.3
mysqlclient==2.1.1
Pillow==9.4.0
psycopg2==2.9.5
PyJWT==2.6.0
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.3
tzdata==2023.3
